// src/platform/plugins/shared/visualizations/public/types.d.ts
import type { CoreStart, SavedObjectsMigrationVersion } from '@kbn/core/public';
import type { IAggConfigs, SerializedSearchSourceFields, TimefilterContract, AggConfigSerialized } from '@kbn/data-plugin/public';
import type { ISearchSource } from '@kbn/data-plugin/common';
import type { SavedObjectsResolveResponse } from '@kbn/core-saved-objects-api-server';
import type { ExpressionAstExpression } from '@kbn/expressions-plugin/public';
import type { TableListTab } from '@kbn/content-management-tabbed-table-list-view';
import type { VisParams } from '@kbn/visualizations-common';
import type { Vis } from './vis';
import type { PersistedState } from './persisted_state';
import type { SerializedVis } from '../common';
export type StartServices = Pick<CoreStart, 'overlays' | 'analytics' | 'i18n' | 'theme' | 'userProfile'>;
export type { Vis, SerializedVis, VisParams };
export interface SavedVisState {
    title: string;
    type: string;
    params: VisParams;
    aggs: AggConfigSerialized[];
}
export interface ISavedVis {
    id?: string;
    title: string;
    description?: string;
    visState: SavedVisState;
    searchSourceFields?: SerializedSearchSourceFields;
    uiStateJSON?: string;
    savedSearchRefName?: string;
    savedSearchId?: string;
    sharingSavedObjectProps?: {
        outcome?: SavedObjectsResolveResponse['outcome'];
        aliasTargetId?: SavedObjectsResolveResponse['alias_target_id'];
        aliasPurpose?: SavedObjectsResolveResponse['alias_purpose'];
        errorJSON?: string;
    };
}
export interface VisSavedObject extends ISavedVis {
    lastSavedTitle: string;
    getEsType: () => string;
    getDisplayName: () => string;
    displayName: string;
    migrationVersion?: SavedObjectsMigrationVersion;
    searchSource?: ISearchSource;
    version?: string;
    tags?: string[];
    managed: boolean;
}
export interface SaveVisOptions {
    confirmOverwrite?: boolean;
    copyOnSave?: boolean;
}
export interface GetVisOptions {
    id?: string;
    searchSource?: boolean;
    migrationVersion?: SavedObjectsMigrationVersion;
    savedSearchId?: string;
    type?: string;
    indexPattern?: string;
}
export interface VisToExpressionAstParams {
    timefilter: TimefilterContract;
    timeRange?: any;
    abortSignal?: AbortSignal;
}
export type VisToExpressionAst<TVisParams extends VisParams = VisParams> = (vis: Vis<TVisParams>, params: VisToExpressionAstParams) => Promise<ExpressionAstExpression> | ExpressionAstExpression;
export interface VisEditorOptionsProps<VisParamType = unknown> {
    aggs: IAggConfigs;
    hasHistogramAgg: boolean;
    isTabSelected: boolean;
    stateParams: VisParamType;
    vis: Vis;
    uiState: PersistedState;
    setValue<T extends keyof VisParamType>(paramName: T, value: VisParamType[T]): void;
    setValidity(isValid: boolean): void;
    setTouched(isTouched: boolean): void;
}
export type ListingViewRegistry = Pick<Set<TableListTab>, 'add'>;